+++ src/constants/settings.ts
@@ -66,6 +66,7 @@ export const SETTING_MEDIA_SUPPORTED_MIMETYPES = "media.supportedMimeTypes";
 
 export const SETTING_DASHBOARD_OPENONSTART = "dashboard.openOnStart";
 export const SETTING_DASHBOARD_CONTENT_TAGS = "dashboard.content.cardTags";
+export const SETTING_DASHBOARD_CONTENT_PAGINATION = "dashboard.content.pagination";
 
 export const SETTING_DATA_FILES = "data.files";
 export const SETTING_DATA_FOLDERS = "data.folders";